bewilson/ci cd (#6)
* first attempt at one of these CI/CD things

* Refactor GitHub Actions workflow for AlmaLinux 9

Updated the GitHub Actions workflow to use AlmaLinux 9 and added steps for installing CVMFS and verifying the mount. Adjusted build steps and ensured the environment is loaded correctly.

* Simplify CVMFS setup using GitHub Action

Replaced manual CVMFS installation steps with a GitHub Action.

* Refactor CI build workflow for improved clarity

* Change CI build branch to 'bewilson/ci-cd/'

* Modify CI workflow to change container and comment steps

Updated the CI build workflow to use a different container image and commented out the installation and mounting steps.

* Change Docker image in GitHub Actions workflow

* try this

* now try this